+++ src/clj/clj_toolbox/base64.clj
@@ -2,7 +2,15 @@
   (:import
     [java.util Base64]))
 
-(defn encode-str
+(defn ^String
+  encode-str
   "base64 encodes the input string. Returns a string"
   [^String s]
   (.encodeToString (Base64/getEncoder) (.getBytes s)))
+
+(defn ^String
+  decode-str
+  [^String s]
+  (-> (Base64/getDecoder)
+      (.decode s)
+      (String.)))